I moved the sections to templates
