+++ src/templates/resume/_home/components/InteractiveContent.js
